feat: 获取插件信息/缓存操作/获取账号信息/消息发送与撤回

+++ http/verify.go
@@ -28,7 +28,7 @@ func Verify(verifyKey string) (result model.VerifyResult) {
 // 使用此方法校验并激活你的Session，同时将Session与一个已登录的Bot绑定
 // 用法同官方文档，错误代码-10000为自定义错误， 错误信息见msg字段
 // NOTE: https://docs.mirai.mamoe.net/mirai-api-http/adapter/HttpAdapter.html#%E7%BB%91%E5%AE%9A
-func Bind(sessionKey string, qq uint64) (result BindResult) {
+func Bind(sessionKey string, qq int) (result BindResult) {
 	const url = "/bind"
 	body := map[string]any{
 		"sessionKey": sessionKey,
@@ -47,7 +47,7 @@ func Bind(sessionKey string, qq uint64) (result BindResult) {
 // 使用此方式释放session及其相关资源（Bot不会被释放） 不使用的Session应当被释放，长时间（30分钟）未使用的Session将自动释放，否则Session持续保存Bot收到的消息，将会导致内存泄露
 // 用法同官方文档，错误代码-10000为自定义错误， 错误信息见msg字段
 // NOTE: https://docs.mirai.mamoe.net/mirai-api-http/adapter/HttpAdapter.html#%E9%87%8A%E6%94%BE
-func Release(sessionKey string, qq uint64) (result ReleaseResult) {
+func Release(sessionKey string, qq int) (result ReleaseResult) {
 	const url = "/release"
 	body := map[string]any{
 		"sessionKey": sessionKey,
